[#9] ViewModel 추가
